swap global var for a func
