+++ types/types.go
@@ -419,19 +419,3 @@ func (h *HealthcheckConfig) GetTimeoutDuration() time.Duration {
 func (h *HealthcheckConfig) GetStartPeriodDuration() time.Duration {
 	return time.Duration(h.StartPeriod) * time.Second
 }
-
-// WaitForPhases we need to init Waitgroups for all the states,
-// this is a helper var listing all the states for initialization.
-var WaitForPhases []WaitForPhase = []WaitForPhase{
-	WaitForCreate, WaitForCreateLinks,
-	WaitForConfigure, WaitForHealthy, WaitForExit,
-}
-
-func WaitForPhaseFromString(s string) (WaitForPhase, error) {
-	for _, val := range WaitForPhases {
-		if s == string(val) {
-			return val, nil
-		}
-	}
-	return "", fmt.Errorf("unknown phase %q", s)
-}
